#19 - a classe "PetEndpoint" para "PetController"
